feat(BREAKING): Manage official containers
The `promasu` containers have not been updated and are deprecated. Use
the official containers now that they are published.

This means that from now on, nginx will need to be managed separately
from the cryptpad container. This formula does this via a simple nginx
container and a variation of the basic nginx config found in the
cryptpad repo.